Fix Firefox
FF42 introduced a native fetch implementation which adds headers beyond those described explicitly. This results in Firefox providing `Accept: text/html` which `express-graphql` can interpret as a desire to return GraphiQL instead of a query result. This adds an `Accept` header explicitly to ensure JSON is returned.

In addition, Firefox often produces errors with no stack, which causes GraphiQL to fall back on using the error itself in displaying messages, however previously we were not casting these to strings, which resulted in errors when string methods were used on them, obfuscating the root issue.

This fixes #65, but a follow-up diff on `express-graphql` will fix any live issuses.