Enable multiple queries
This enables GraphiQL to support multiple queries in the editor pane at the same time. When multiple queries exist, the run button becomes a drop-down menu showing the names of all queries available to run - forcing a selection. Cmd+Enter still runs a query, but it runs whatever the last query was unless the cursor is within the bounds of a query - then it runs that query.

This implements #81 and #40.