Use autoprefixer
Follow-up to 8b4fe782c3d713ce. We shouldn't have to be manually
micro-managing our CSS prefixes, which is an error prone process.

Instead, let's use autoprefixer. This commit sets that up and also
removes the now-unnecessary prefixes from the source CSS. To get a sense
of what's changed, here are the diffs between:

- the original output; and:
- the original source files piped through autoprefixer:

  https://gist.github.com/wincent/bdc72e085b97158e4cc87e8c1226dd25

- the original output; and:
- the stripped down source files piped thorugh autoprefixer:

  https://gist.github.com/wincent/bd41cb2d0cb92c6b77eec74d02db0798

Exceptions: I didn't remove the `-o-ellipsis-lastline` prefix nor the
`-webkit-tap-highlight-color` one, because those aren't handled by
autoprefixer (it only deals with stuff in the caniuse.com tables).

Tested in Chrome and Firefox so far, but will test a couple more
browsers while this is under review.