Fix Flow error
This one was a legitimate mis-typing, so I am not sure why Flow used to
be blind to the error when this exact same file lived two levels higher
in the repo.

Oh well. ¯\_(ツ)_/¯

Confirmed that the sole caller of `validateWithCustomRules` is indeed
passing in a `DocumentNode` and not `ASTNode`, because the value is
coming from `parse` in the `graphql` package (which returns a
`DocumentNode`).