Shorten a line to avoid problem with prettier
If this line is too long, prettier wants to wrap this call across
multiple lines, and it wants to append a trailing comma too. That's fine
on our Babel-transpiled files, but is no good on this one, which needs
to be acceptable by vanilla node. So, rather than making an ugly
override comment, let's just reword the error to fit all on one line.