Make main package use standard build-js.js script
Now the main package uses the same build technique as the subpackages.
Tested this by doing a before and after diff of the "dist" directory;
the generated files are exactly the same, except for the package.json,
which we don't need to redundantly copy to the dist directory (this was
copied from another GraphQL project where we edit the package.json file
programmatically during the Travis build for the purposes of cutting a
release).