Add more specific types
I think this is a net improvement; drops us from 222 to 143 errors. (It
is hard to be entirely certain: given the size of the change it may be
spurious, but it is hard to be sure because the error output is still so
noisy. Nevertheless, looks like a relatively "safe" change because it is
a type-only change with no code changes.)