Remove broken top-level build scripts
These are all broken because they assume the difference of "src" at the
top level, which ceased to exist in 3b93477ed2c20bbfe6b8fa47460.

Note: other stuff on the top-level is still broken, like "testonly" and
therefore "test" as well, which in turn means that Travis CI is broken
(and would be anyway because top-level Flow "check" is broken; I haven't
tested Flow inside "packages/*" yet).